Created calculation “setANewWeight”. Make sure to check out the pre-conditions on this function. It edits the preference Manager. Rewrote setWeight to not cause infinite loop. Rewrote weights throughout app to be "float", not "int"

+++ College Rank/UserPreference.h
@@ -25,7 +25,7 @@
 -(void) setWeight:(float)wt;
 -(void) changeLock;
 -(bool) getLock;
--(int) getWeight;
+-(float) getWeight;
 -(int) getPrefVal;
 -(void) setPrefVal: (int) prf;
 -(NSString*) getName;
